+++ linpg/core/__init__.py
@@ -1,6 +1,6 @@
 """
 结构:
-system -> module -> shape -> surface -> text -> image -> frame -> window -> video -> sprite
+system -> module -> shape -> surface -> text -> image -> window -> video -> sprite
 """
 
 from .sprite import *
